Add delete Experience functionality

+++ src/app/experiences/experiences.component.ts
@@ -23,7 +23,7 @@ export class ExperiencesComponent implements OnInit {
 
   ngOnInit() {
     this.experienceForm = new FormGroup({
-      experienceName: new FormControl(''),      
+      experienceName: new FormControl(''),
       experienceLoginUrl: new FormControl('http://<IP_ADDRESS>:8080/gateway/api/auth/login'),
       experienceModelUrl: new FormControl('http://<IP_ADDRESS>:8080/gateway/api/portals'),
     });
@@ -32,23 +32,25 @@ export class ExperiencesComponent implements OnInit {
     this.experiences$ = this.store.pipe(select(experiencesSelectors.getExperiences));
   }
 
-  deleteExperience(experienceName: string) {
-    this.store.dispatch(new experiencesActions.DeleteExperience(experienceName));
+  deleteExperience(experienceName: string, event) {
+    if (confirm('Are you sure to delete')) {
+      this.store.dispatch(new experiencesActions.DeleteExperience(experienceName));
+    }
+    event.stopPropagation();
     return false;
   }
 
   openAddExperienceModal(addExperienceFormContent) {
-    this.modalService.open(addExperienceFormContent, { ariaLabelledBy: 'modal-basic-title' }).result.then((result) => {}, (reason) => {});
+    this.modalService.open(addExperienceFormContent, { ariaLabelledBy: 'modal-basic-title' }).result.then((result) => { }, (reason) => { });
   }
 
   addExperience() {
-    let experienceFormValue =  this.experienceForm.value;
+    let experienceFormValue = this.experienceForm.value;
 
     this.store.dispatch(new experiencesActions.SyncExperience({
-      portalName: experienceFormValue.experienceName,      
+      portalName: experienceFormValue.experienceName,
       loginUrl: experienceFormValue.experienceLoginUrl,
       portalUrl: experienceFormValue.experienceModelUrl,
     }));
-    //this.service.syncModel(this.portalForm.value)
   }
 }
